+++ src/srgp/srgplocal.h
@@ -284,7 +284,7 @@ void SRGP__initColor (int requested_planes);
 void SRGP__cleanupColor (void);
 void SRGP__activateApplColorTable (void);
 void SRGP__deactivateApplColorTable (void);
-unsigned short SRGP__retrieveColorFromPallette(int colorIndex);
+unsigned long SRGP__retrieveColorFromPallette(int colorIndex);
